Drop support for all HITANS segments except v6
PSD2 specifically requires HITANSv6, all banks support it in practice. Older segments are only useful to interact with banks that only have accounts not affected by PSD2 (stock depots, perhaps some business accounts) and therefore do not need to upgrade to HITANSv6. I don't know if such banks even exist, but if so, supporting them would additionally require the older HKTAN and HITAN segments in this library anyway.
