--- lib/Fhp/Segment/HITANS/HITANS.php
@@ -1,16 +1,0 @@
-<?php
-
-namespace Fhp\Segment\HITANS;
-
-use Fhp\Segment\SegmentInterface;
-
-/**
- * Segment: Zwei-Schritt-TAN-Einreichung, Parameter
- * Bezugssegment: HKVVB
- * Sender: Kreditinstitut
- */
-interface HITANS extends SegmentInterface
-{
-    /** @return ParameterZweiSchrittTanEinreichung */
-    public function getParameterZweiSchrittTanEinreichung(): ParameterZweiSchrittTanEinreichung;
-}
