--- lib/Fhp/Segment/HITANS/HITANSv5.php
@@ -1,25 +1,0 @@
-<?php /** @noinspection PhpUnused */
-
-namespace Fhp\Segment\HITANS;
-
-use Fhp\Segment\BaseGeschaeftsvorfallparameter;
-
-/**
- * Segment: Zwei-Schritt-TAN-Einreichung, Parameter (Version 5)
- * Parameters for: HKTANv5
- * Bezugssegment: HKVVB
- * Sender: Kreditinstitut
- *
- * @link: https://www.hbci-zka.de/dokumente/spezifikation_deutsch/archiv/FinTS_V3.0_2017-10-06-FV_RM.zip
- */
-class HITANSv5 extends BaseGeschaeftsvorfallparameter implements HITANS
-{
-    /** @var ParameterZweiSchrittTanEinreichungV5 */
-    public $parameterZweiSchrittTanEinreichung;
-
-    /** @return ParameterZweiSchrittTanEinreichungV5 */
-    public function getParameterZweiSchrittTanEinreichung(): ParameterZweiSchrittTanEinreichung
-    {
-        return $this->parameterZweiSchrittTanEinreichung;
-    }
-}
