--- lib/Fhp/Segment/HITANS/ParameterZweiSchrittTanEinreichung.php
@@ -1,12 +1,0 @@
-<?php /** @noinspection PhpUnused */
-
-namespace Fhp\Segment\HITANS;
-
-interface ParameterZweiSchrittTanEinreichung
-{
-    /** @return bool */
-    public function getEinschrittVerfahrenErlaubt(): bool;
-
-    /** @return VerfahrensparameterZweiSchrittVerfahren[] */
-    public function getVerfahrensparameterZweiSchrittVerfahren(): array;
-}
