--- lib/Fhp/Segment/HITANS/ParameterZweiSchrittTanEinreichungV4.php
@@ -1,34 +1,0 @@
-<?php /** @noinspection PhpUnused */
-
-namespace Fhp\Segment\HITANS;
-
-use Fhp\Segment\BaseDeg;
-
-class ParameterZweiSchrittTanEinreichungV4 extends BaseDeg implements ParameterZweiSchrittTanEinreichung
-{
-    /** @var bool */
-    public $einschrittVerfahrenErlaubt;
-    /** @var bool */
-    public $mehrAlsEinTanPflichtigerAuftragProNachrichtErlaubt;
-    /**
-     * 0: Auftrags-Hashwert nicht unterstützt
-     * 1: RIPEMD-160
-     * 2: SHA-1
-     * @var int
-     */
-    public $auftragsHashwertverfahren;
-    /** @var VerfahrensparameterZweiSchrittVerfahrenV4[] @Max(98) */
-    public $verfahrensparameterZweiSchrittVerfahren;
-
-    /** @return bool */
-    public function getEinschrittVerfahrenErlaubt(): bool
-    {
-        return $this->einschrittVerfahrenErlaubt;
-    }
-
-    /** @return VerfahrensparameterZweiSchrittVerfahrenV4[] */
-    public function getVerfahrensparameterZweiSchrittVerfahren(): array
-    {
-        return $this->verfahrensparameterZweiSchrittVerfahren;
-    }
-}
