+++ lib/Fhp/Segment/TAN/HITANSv6.php
@@ -1,6 +1,6 @@
 <?php /** @noinspection PhpUnused */
 
-namespace Fhp\Segment\HITANS;
+namespace Fhp\Segment\TAN;
 
 use Fhp\Segment\BaseGeschaeftsvorfallparameter;
 
@@ -12,17 +12,9 @@ use Fhp\Segment\BaseGeschaeftsvorfallparameter;
  *
  * @link: https://www.hbci-zka.de/dokumente/spezifikation_deutsch/fintsv3/FinTS_3.0_Security_Sicherheitsverfahren_PINTAN_2018-02-23_final_version.pdf
  * Section: B.5.1 c)
- *
- * TODO Move HITANS to the TAN namespace. Probably we can drop all HITANS older than v6.
  */
-class HITANSv6 extends BaseGeschaeftsvorfallparameter implements HITANS
+class HITANSv6 extends BaseGeschaeftsvorfallparameter
 {
     /** @var ParameterZweiSchrittTanEinreichungV6 */
     public $parameterZweiSchrittTanEinreichung;
-
-    /** @return ParameterZweiSchrittTanEinreichungV6 */
-    public function getParameterZweiSchrittTanEinreichung(): ParameterZweiSchrittTanEinreichung
-    {
-        return $this->parameterZweiSchrittTanEinreichung;
-    }
 }
